# perf(nebula): optimize AT128 decoder performance