--- a/nebula_decoders/include/nebula_decoders/nebula_decoders_hesai/decoders/pandar_at_decoder.hpp
+++ b/nebula_decoders/include/nebula_decoders/nebula_decoders_hesai/decoders/pandar_at_decoder.hpp
@@ -32,7 +32,7 @@ const uint16_t LIDAR_AZIMUTH_UNIT = 256;
 const uint32_t MAX_AZI_LEN = 36000 * 256;
 
 /// @brief The resolution of one scan (in single return mode)
-const uint16_t SCAN_POINTS_NUM = 1200 * LASER_COUNT;
+const uint32_t SCAN_POINTS_NUM = 1200 * LASER_COUNT;
 /// @brief This sensor supports a maximum of 2 returns (= dual return)
 const uint16_t MAX_RETURN_COUNT = 2;