--- a/nebula_ros/include/nebula_ros/hesai/hesai_decoder_ros_wrapper.hpp
+++ b/nebula_ros/include/nebula_ros/hesai/hesai_decoder_ros_wrapper.hpp
@@ -13,6 +13,8 @@
 #include <rclcpp/rclcpp.hpp>
 #include <rclcpp_components/register_node_macro.hpp>
 
+#include <chrono>
+
 #include "pandar_msgs/msg/pandar_packet.hpp"
 #include "pandar_msgs/msg/pandar_scan.hpp"