## PR Type

- Improvement

## Related Links

--

## Description

This PR improves performance of the Hesai AT128 decoder by approx. 23% without changing the logical output.
To achieve this, three changes have been made:
* Decoding points directly to `scan_pc_` and `overflow_pc_` instead of temporary buffers and re-using the buffers (~ 11%point improvement)
* Removing unused copy operation for incoming scan packets (~ 9%point improvement)
* Combining azimuth and elevation sin/cos lookup tables as they were exact duplicates (~ 4%point improvement)

Points were decoded in `convert(block_id)` to a temporary PointCloud, the contents of which were then copied to `scan_pc_` or `overflow_pc_` respectively. This caused avoidable latency. Now, depending on whether the current block being decoded is in a new frame/scan or not, `convert(block_id, pcl)` decodes those points directly into the right buffer.

Further, `overflow_pc_` was reallocated on each full scan, causing avoidable heap allocations.
Because each incoming `PandarScan` message always contains a full scan (albeit not necessarily aligned to our phase setting), `scan_decoder->get_pointcloud()` will be called exactly once in `ConvertScanToPointCloud`. Thus, it is guaranteed that one of `scan_pc_` and `overflow_pc_` can always be used for decoding, and the other as the output buffer.
Once a scan is complete, the buffers are swapped and the `overflow_pc_` is `clear`ed (not reallocated).

No changes have been made to the base class, the new function is an overloaded version of the existing one (which has been changed to throw a "not implemented" error).

![f7249549-193f-4193-a43a-ef47260f2416](https://github.com/tier4/nebula/assets/6088931/0bf4cdb1-5a3a-4d75-9e75-017a8412c6b7)

|                     |Before|Optimized|
|-----------------|---------|--------------|
|duration AVG        |7.972725 ms      |6.147384 ms|
|duration STD         |0.759695       |0.613904|
|duration AVG % rel. to before        |100.000000      | **77.105188** |
|duration STD % rel. to before    |100.000000      |80.809372|

All measurements have been made on an Intel Core i7-12700H with
* TurboBoost disabled
* All cores set to 2.3 GHz
* Thread pair 2/3 isolated
* Nebula `taskset` to core 2

Measurements were repeated 3 times and ran for 37s each (~370 iterations).

## Review Procedure

To confirm logical equality with the previous implementation, a rosbag of recorded AT128 packets can be used.
To confirm performance improvement, I provide the scripts and logging that yielded the above charts:
```
scripts/
    test_runner.bash  # runs the driver for 3 iterations on a given rosbag
    plot_times.py     # takes the logs from test_runner.bash and creates a plot like the one above
```
The test runner locks/unlocks core frequencies and `taskset`s nebula onto the specified core(s). Ideally, the taskset cores are isolated (cf. [Autoware performance measurement guide](https://autowarefoundation.github.io/autoware-documentation/main/how-to-guides/integrating-autoware/tuning-parameters-and-performance/evaluating-real-time-performance/#prepare-separated-cores))

## Remarks

<!-- Write remarks as you like if you need them. -->

## Pre-Review Checklist for the PR Author

*PR Author should check the checkboxes below when creating the PR.*
- [x] Assign PR to reviewer

## Checklist for the PR Reviewer

*Reviewers should check the checkboxes below before approval.*
- [x] Commits are properly organized and messages are according to the guideline
- [x] (Optional) Unit tests have been written for new behavior
- [x] PR title describes the changes

## Post-Review Checklist for the PR Author

*PR Author should check the checkboxes below before merging.*
- [x] All open points are addressed and tracked via issues or tickets

## CI Checks

- *Build and test for PR*: Required to pass before the merge.