@amc-nu 2023-08-08
@mojomex can you please attend to the spelling errors on the [spell-check job](https://github.com/tier4/nebula/actions/runs/5785120036/job/15696013740?pr=37)?