@mojomex 2023-08-08
Most of the warnings were the names of Bash commands, added them to the dictionary. `stds` for standard deviations is also commonly used in Python/Pandas, added to the dictionary as well.